% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_own_functions.R
\name{check_company_ard_function}
\alias{check_company_ard_function}
\title{Try one of the company's ARD functions}
\usage{
check_company_ard_function(
  name,
  data,
  ...,
  stat_names = NULL,
  home = tflplanner_home()
)
}
\arguments{
\item{name}{The function.}

\item{data}{The data to try it on.}

\item{...}{Its other arguments (\verb{by = TRT01A, variables = AGE}).}

\item{stat_names}{The statistics it should give.}

\item{home}{The tflplanner home.}
}
\value{
The problems found (\code{\link[tflspec:tfl_check_ard_function]{tflspec::tfl_check_ard_function()}}).
}
\description{
Loads its file and runs \code{\link[tflspec:tfl_check_ard_function]{tflspec::tfl_check_ard_function()}} on \code{data},
with the arguments an analysis row would give.
}
