% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study_files}
\alias{study_files}
\title{Files in a study folder}
\usage{
study_files(study, folder = "data")
}
\arguments{
\item{study}{An \code{rtfstudy}.}

\item{folder}{A folder of \code{\link[=study_layout]{study_layout()}} (\code{"adam"}, \code{"tfl"}, ...), or
\code{"data"} for all three data folders.}
}
\value{
A data frame: \code{folder}, \code{file}, \code{size_kb}, \code{modified}, \code{path}.
}
\description{
Files in a study folder
}
