% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study_layout}
\alias{study_layout}
\title{The folder layout of a study}
\usage{
study_layout()
}
\value{
A named character vector of folders, relative to the study
folder.
}
\description{
The folder layout of a study
}
