% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home.R
\name{tflplanner_config}
\alias{tflplanner_config}
\alias{studies_root}
\title{tflplanner's settings}
\usage{
tflplanner_config(home = tflplanner_home())

studies_root(home = tflplanner_home())
}
\arguments{
\item{home}{The home folder.}
}
\value{
A list: \code{studies_root}, \code{last_study}, \code{language}.
}
\description{
tflplanner's settings
}
