% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_spec.R
\name{read_ard_spec}
\alias{read_ard_spec}
\alias{ard_spec}
\title{Read and check an ARD definition workbook}
\usage{
read_ard_spec(path)

ard_spec(x)
}
\arguments{
\item{path}{An \code{ard_spec.xlsx}.}

\item{x}{A list of the sheets (data frames).}
}
\value{
An \code{ard_spec}: a list of the four sheets.
}
\description{
Read and check an ARD definition workbook
}
