% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home.R
\name{rtfplanner_config}
\alias{rtfplanner_config}
\alias{studies_root}
\title{rtfplanner's settings}
\usage{
rtfplanner_config(home = rtfplanner_home())

studies_root(home = rtfplanner_home())
}
\arguments{
\item{home}{The home folder.}
}
\value{
A list: \code{studies_root}, \code{last_study}.
}
\description{
rtfplanner's settings
}
