% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflplanner-package.R
\docType{package}
\name{tflplanner-package}
\alias{tflplanner}
\alias{tflplanner-package}
\title{tflplanner: a Shiny study manager for clinical TFLs}
\description{
The GUI over tflspec (the Excel specifications, ARD and plan engine) and
rtfreporter (the RTF renderer).
}
\section{Start here}{

\code{setup_tflplanner()} once: where the app keeps its settings and where
new study folders go (its home, under \code{\link[tools:userdir]{tools::R_user_dir()}} unless you
choose another).  Then \code{run_app()} opens the app.  \code{create_sample_study()}
makes a study to try it on: CDISC pilot data with tables, a listing
and figures, ready to run.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ichirio/tflplanner}
  \item \url{https://ichirio.github.io/tflplanner/}
  \item Report bugs at \url{https://github.com/ichirio/tflplanner/issues}
}

}
\author{
\strong{Maintainer}: Yoichi Masui \email{ichi.r.pro@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Cytel Inc. (pharmaverseadam data in inst/sample (Apache License 2.0)) [copyright holder]
  \item F. Hoffmann-La Roche AG (pharmaverseadam data in inst/sample (Apache License 2.0)) [copyright holder]
  \item GlaxoSmithKline LLC (pharmaverseadam data in inst/sample (Apache License 2.0)) [copyright holder]
}

}
\keyword{internal}
