% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflplanner-package.R
\docType{package}
\name{tflplanner-package}
\alias{tflplanner}
\alias{tflplanner-package}
\title{tflplanner: a Shiny study manager for clinical TFLs}
\description{
The GUI over tflspec (the Excel specifications, ARD and plan engine) and
rtfreporter (the RTF renderer).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ichirio/tflplanner}
  \item \url{https://ichirio.github.io/tflplanner/}
  \item Report bugs at \url{https://github.com/ichirio/tflplanner/issues}
}

}
\author{
\strong{Maintainer}: Yoichi Masui \email{ichi.r.pro@gmail.com} [copyright holder]

}
\keyword{internal}
